requirements.txt
fastapi
uvicorn[standard]
openai
httpx
supabase
python-dotenv
python-multipart
paddleocr
paddlepaddle
pdfplumber
aiofiles
